[hotfix][runtime] Refactor testing methods out from SlotProfile into SlotProfileTestingUtils
